Add a Python3Target class to the tool.
(It's a copy of PythonTarget, but with s/PythonTarget/Python3Target/.)

This is necessary so that generated code correctly compares the bytes
(given by LA) to character ordinals, for literal tokens.

Also add an explicit cast in GrammarAST, to shut javac up so I can
compile the Python3Target into a class for testing.