Convert dfaEdgeSwitch to use set inclusion.
This makes it so the condition is eg. "LA3 in {FOO, BAR, BAZ}"
instead of "LA3 == FOO or LA3 == BAR or LA3 == BAZ".

Unfortunately other cases that would benefit from this approach
(such as setTest) are structured in such a way that they don't have
an easy way (so far as I can tell) to discern what to be checking
("LA3" in the above example).