Allow genSetExpr/setTest to actually use sets.
This adds three additional possible arguments to setTest which allow
language templates to use set inclusion tests instead of repeated
equality and range checking tests:
- varName, which is either the temp variable the LA token is saved to,
  or the LA() function call. This is given by two new templates,
  lookaheadVarName and isolatedLookaheadVarName, which match their
  counterpart lookaheadTest rules.
- allValues, the flattened full list of token names to match against.
- allValuesAsInt, the flattened full list of token ids to match against,
  similar to atomAsInt in the lookaheadTest rules.

To receive allValues and/or allValuesAsInt, all three of these arguments
must be listed in the setTest definition. This also stops the ranges
argument from being passed in. Alternately, the setTest template can define
only ranges and varName as arguments, in which case it does not receive
allValues or allValuesAsInt. (This can be useful if you'd like to test
set inclusion for only the individual elements in ranges, and keep the
range tests themselves.)

And finally, make use of this feature in the Python3 target, since
Python3 allows raw definitions of sets, making these checks consistently
faster than the old behavior.