Fix setTest making extraordinarily large ranges.
This changes how a language tests for set inclusion.
Instead of changing the input parameters to the setTest template,
define templates lookaheadSetTest and isolatedLookaheadSetTest.
These will be like their counterparts for atom and range, but will
have a list of values to use.

The generated code from this SetTest will be treated as a range--i.e.
it will be joined in setTest with the other ranges, where, in Python3
for example, it will be one of the conditions 'or'ed together.

This commit specifies that that set will consist of all the values in
the provided IntervalSet that appeared in an Interval no larger than 4.
Thus, this accomplishes a strict improvement over most normal cases,
without blowing up for cases where the IntervalSet contains all of
Unicode save for a few characters.