Remove const specifiers from CyclicDFA members
These const specifiers were ignored by older versions of gcc, likely
because they are in template members that are not instantiated. gcc 14
refuses to compile the class.

Fix by removing the const specifiers. This is safe since these members
are private, so we aren't allowing external users permissions to
modify them.